+++ extensions/dou-tu/src/index.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Grid } from "@raycast/api";
+import { Action, ActionPanel, getPreferenceValues, Grid } from "@raycast/api";
 import clipboardService from "./services/clipboardService";
 import sourcesService from "./services/sourcesService";
 import { IDoutuImage } from "./services/sources";
@@ -12,16 +12,22 @@ let awaitRequest = false;
 export default function Command() {
   const [isLoading, setIsLoading] = useState(true);
   const [isEnd, setIsEnd] = useState(true);
-  const [selectedItemId, setSelectedItemId] = useState("placeholder_1");
   const [list, setList] = useState<IDoutuImage[]>([]);
+  const [selectMode, setSelectMode] = useState("actions");
+  // eslint-disable-next-line prefer-const
+  let [selectedItemId, setSelectedItemId] = useState<string | undefined>();
 
   useEffect(() => {
     more();
+    const { select_mode } = getPreferenceValues<{
+      select_mode: string;
+    }>();
+    setSelectMode(select_mode);
+    if (select_mode === "click") setSelectedItemId("placeholder_1");
   }, []);
 
   const more = async () => {
     setIsLoading(true);
-    awaitRequest = true;
     const placeholderItem = { id: `placeholder_${currentPageIndex}_${uuidv4()}`, url: "" };
     const result = await sourcesService.get(currentKeyword, currentPageIndex++);
     setIsEnd(result.isEnd);
@@ -29,10 +35,19 @@ export default function Command() {
       currentPageIndex = -1;
       setList([]);
     } else {
-      setList([...(currentPageIndex === 2 ? [] : list), placeholderItem, ...result.images]);
-      setSelectedItemId(placeholderItem.id);
+      if (selectMode === "click") {
+        awaitRequest = true;
+        setList([...(currentPageIndex === 2 ? [] : list), placeholderItem, ...result.images]);
+        setSelectedItemId(placeholderItem.id);
+      } else setList([...(currentPageIndex === 2 ? [] : list), ...result.images]);
     }
     setIsLoading(false);
+  };
+
+  const selectAction = (id: string | undefined) => {
+    if (!id) return;
+    const item = list.find((o) => o.id === id);
+    item && clipboardService.imageToClipboard(item.url);
   };
 
   const searchBarAccessory = (
@@ -54,6 +69,12 @@ export default function Command() {
     </Grid.Dropdown>
   );
 
+  const actions = (
+    <ActionPanel>
+      {selectMode === "actions" ? <Action title="Copy Image" onAction={() => selectAction(selectedItemId)} /> : null}
+    </ActionPanel>
+  );
+
   return (
     <Grid
       throttle={true}
@@ -64,19 +85,17 @@ export default function Command() {
         currentPageIndex = 1;
         more();
       }}
-      onSelectionChange={async (id) => {
+      onSelectionChange={(id: string | undefined) => {
         if (!id) return;
-        if (id.startsWith("placeholder_")) {
-          awaitRequest = false;
-          return;
+        if (selectMode === "click") {
+          if (awaitRequest) return (awaitRequest = !id.startsWith("placeholder_"));
+          if (id.startsWith("placeholder_")) return;
+          if (id === "more" && currentPageIndex > 0) return more();
+          selectAction(id);
+        } else {
+          if (id === "more" && currentPageIndex > 0) return more();
+          selectedItemId = id;
         }
-        if (awaitRequest) return;
-        if (id === "more") {
-          currentPageIndex > 0 && more();
-          return;
-        }
-        const item = list.find((o) => o.id === id);
-        item && clipboardService.imageToClipboard(item.url);
       }}
       searchBarAccessory={searchBarAccessory}
     >
@@ -86,20 +105,20 @@ export default function Command() {
             key={index}
             id={item.id.toString()}
             content={{
-              tooltip: "Click images copy",
+              tooltip: "Click Images Copy",
               value: {
                 source: "click.png",
               },
             }}
           />
         ) : (
-          <Grid.Item key={index} id={item.id.toString()} content={{ source: item.url }} />
+          <Grid.Item key={index} id={item.id.toString()} content={{ source: item.url }} actions={actions} />
         )
       )}
       {isEnd ? (
         <></>
       ) : (
-        <Grid.Item key="more" id="more" content={{ tooltip: "Click more", value: { source: "more.png" } }} />
+        <Grid.Item key="more" id="more" content={{ tooltip: "Click More", value: { source: "more.png" } }} />
       )}
     </Grid>
   );