+++ extensions/dou-tu/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 斗图 Changelog
 
-## [Add Dou Tu Extension] - 2022-08-31
+## [Add Dou Tu Extension] - 2022-09-12
+
 - Support for quick search of emoji packs
 - Support click to copy to clipboard
+- Support enter key to copy to clipboard
+- Support toggle image sources